requirements.txt
djangorestframework>=3.15.2
pytest-django==4.9.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
